updating auth command to work with new routing
