+++ .tode/.template/auth/route/routes.txt
@@ -1,0 +1,2 @@
+App.post('/auth/register', loadController('auth.store'));
+App.post('/auth/login', loadController('auth.login'));
